[Refactor] Move Constants to base package
