+++ src/main/java/info/freelibrary/iiif/presentation/AnnotationList.java
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.annotation.JsonSetter;
 
 import info.freelibrary.iiif.presentation.properties.Behavior;
 import info.freelibrary.iiif.presentation.properties.behaviors.ResourceBehavior;
-import info.freelibrary.iiif.presentation.utils.Constants;
 
 /**
  * An ordered list of annotations, typically associated with a single canvas.
