+++ src/main/java/info/freelibrary/iiif/presentation/properties/selectors/ImageApiSelector.java
@@ -1,7 +1,7 @@
 
 package info.freelibrary.iiif.presentation.properties.selectors;
 
-import info.freelibrary.iiif.presentation.utils.Constants;
+import info.freelibrary.iiif.presentation.Constants;
 import info.freelibrary.iiif.presentation.utils.MessageCodes;
 import info.freelibrary.util.Logger;
 import info.freelibrary.util.LoggerFactory;
