+++ src/main/java/info/freelibrary/iiif/presentation/services/ImageInfoService.java
@@ -6,7 +6,7 @@ import java.net.URI;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-import info.freelibrary.iiif.presentation.utils.Constants;
+import info.freelibrary.iiif.presentation.Constants;
 
 /**
  * A service that will return information about a particular image.
