+++ src/test/java/info/freelibrary/iiif/presentation/properties/I18nTest.java
@@ -14,7 +14,7 @@ import org.junit.Test;
 
 import com.thedeanda.lorem.LoremIpsum;
 
-import info.freelibrary.iiif.presentation.utils.Constants;
+import info.freelibrary.iiif.presentation.Constants;
 import info.freelibrary.iiif.presentation.utils.MessageCodes;
 import info.freelibrary.util.Logger;
 import info.freelibrary.util.LoggerFactory;
