+++ src/test/java/info/freelibrary/iiif/presentation/properties/selectors/SelectorDeserializerTest.java
@@ -7,7 +7,7 @@ import static org.junit.Assert.assertNull;
 import org.junit.Test;
 
 import info.freelibrary.iiif.presentation.AbstractTest;
-import info.freelibrary.iiif.presentation.utils.Constants;
+import info.freelibrary.iiif.presentation.Constants;
 
 import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
